[Serve] Add triton mock lib
